telegram_bot/requirements.txt
python-telegram-bot==21.9
pymongo==4.6.1
python-dotenv==1.0.0
motor==3.3.2
